# public/docs/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Lumen: Loja de vendas",
        "version": "0.1"
    },
    "paths": {
...
        "/sales/create": {
            "post": {
                "tags": [
                    "Venda"
                ],
                "summary": "Criar uma nova venda",
                "operationId": "7434c8adc87b45691cd100dfc68edda4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "sale_date",
                                    "total_value"
                                ],
                                "properties": {
                                    "sale_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-04-06"
                                    },
                                    "total_value": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 299.9
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Venda criada com sucesso"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                }
            }
        },
        "/sales/{id}": {
            "get": {
                "tags": [
                    "Venda"
                ],
                "summary": "Buscar uma venda pelo ID",
                "operationId": "cceac3d445ec923f95b0dfcb0a464675",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da venda",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Venda encontrada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "sale_date": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2025-04-06"
                                        },
                                        "total_value": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 299.9
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2025-04-06T12:34:56Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2025-04-06T14:00:00Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Venda não encontrada"
                    }
                }
            }
        },
        "/sales/update/{sale_id}": {
            "put": {
                "tags": [
                    "Venda"
                ],
                "summary": "Atualizar uma venda existente",
                "operationId": "6549216fc5f6e56c98d3f0241e759d09",
                "parameters": [
                    {
                        "name": "sale_id",
                        "in": "path",
                        "description": "ID da venda",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "sale_date",
                                    "total_value"
                                ],
                                "properties": {
                                    "sale_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-04-06"
                                    },
                                    "total_value": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 299.9
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Venda atualizada com sucesso"
                    },
                    "404": {
                        "description": "Venda não encontrada"
                    }
                }
            }
        },
        "/sales/delete/{sale_id}": {
            "delete": {
                "tags": [
                    "Venda"
                ],
                "summary": "Deletar uma venda",
                "operationId": "bf90f63a1c9d7aff12dbce43d8614468",
                "parameters": [
                    {
                        "name": "sale_id",
                        "in": "path",
                        "description": "ID da venda",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Venda deletada com sucesso"
                    },
                    "404": {
                        "description": "Venda não encontrada"
                    }
                }
            }
        },
...
    },
...
}